Refactored so endpoint details are managed in-controller
This changes the setup of things so that controllers manage meta-details
like the endpoint category name and the methods provided, rather than
those being managed in other parts of the app.
This prevents the need of managing references of these details across
other files in the app, since these details can be accessed dynamically
and programatically via the controllers.
An Endpoints class handles references to all controllers and the high
level actions to access the information of these.